Relational Database System (RDBMS)
• Designed using JAVA and its external libraries like JSqlParser and
JDBM2.
• Implemented a simple SQL query evaluator with support for Select,
Project, Join, Union, and Aggregate operations. The code was expected to
evaluate the SELECT statements on provided data, and produce output in a
standardized form.
• Cost optimizations such as external sorting and Sort-merge joins were
implemented to reduce the query evaluation time.
• Hash and Tree Indexes were created using JDBM2 to further reduce the
scanning time on the given relations.
• Implemented the use of INSERT, DELETE, and UPDATE statements. A file
containing multiple operations was presented to the parser, and was
expected to fully evaluate all queries and database updates included in
it.